5.3 Usage Examples - code formatting
The code formatting on the examples in
http://www.w3.org/TR/touch-events/#usage-examples is a bit confusing at
the moment, with some weird indentation and some overlay long lines (and
lack of a <script></script> element).

Could these be prettified/rationalised a bit?

(it seems that as part of the transition to github, some of the
problemtic things have been fixed already, like the broken newlines that
made some lines too long...)

https://lists.w3.org/Archives/Public/public-touchevents/2014Oct/0019.html