Add a bit more documentation of and example for touch constructors
This follows the example of (and copies some prose from) the pointer
events spec.

Note that ideally terms like UIEvent and EventModifierInit would link
directly to their definition in other specs, but it doesn't appear to
be easy to get respec to do this, especially within WebIDL
(in contrast to bikeshed which does a great job of that).